# internal/api/user_handlers.go
// internal/api/user_handlers.go
package api

import (
	"encoding/json"
	"errors"
	"fmt"
	"net/http"
	"strings"

	"github.com/charmbracelet/log"
	"github.com/gin-gonic/gin"

	"github.com/srl-labs/clab-api-server/internal/auth"
	"github.com/srl-labs/clab-api-server/internal/models"
)

type userUpdatePayload struct {
	models.UserUpdateRequest
	hasGroups      bool
	hasIsSuperuser bool
}

func (p *userUpdatePayload) UnmarshalJSON(data []byte) error {
	type userUpdateRequest models.UserUpdateRequest

	var req userUpdateRequest
	if err := json.Unmarshal(data, &req); err != nil {
		return err
	}

	var fields map[string]json.RawMessage
	if err := json.Unmarshal(data, &fields); err != nil {
		return err
	}

	p.UserUpdateRequest = models.UserUpdateRequest(req)
	for field := range fields {
		switch {
		case strings.EqualFold(field, "groups"):
			p.hasGroups = true
		case strings.EqualFold(field, "isSuperuser"):
			p.hasIsSuperuser = true
		}
	}

	return nil
}

// @Summary List users
// @Description Returns a list of system users. Requires superuser privileges.
// @Tags Users
// @Security BearerAuth
// @Produce json
// @Success 200 {array} models.UserDetails "List of user details"
// @Failure 401 {object} models.ErrorResponse "Unauthorized"
// @Failure 403 {object} models.ErrorResponse "Forbidden (User is not a superuser)"
// @Failure 500 {object} models.ErrorResponse "Internal server error"
// @Router /api/v1/users [get]
func ListUsersHandler(c *gin.Context) {
	username := c.GetString("username")

	// Only superusers can list all users
	if !requireSuperuser(c, username, "list all users") {
		return
	}

	users, err := auth.GetAllUsers()
	if err != nil {
		log.Errorf("Failed to get user list: %v", err)
		c.JSON(http.StatusInternalServerError, models.ErrorResponse{Error: "Failed to retrieve user list: " + err.Error()})
		return
	}

	c.JSON(http.StatusOK, users)
}

// @Summary Get user details
// @Description Returns details for a specific user. Requires superuser privileges or the user's own account.
// @Tags Users
// @Security BearerAuth
// @Produce json
// @Param username path string true "Username to get details for" example="john"
// @Success 200 {object} models.UserDetails "User details"
// @Failure 401 {object} models.ErrorResponse "Unauthorized"
// @Failure 403 {object} models.ErrorResponse "Forbidden (Not superuser or not the user's own account)"
// @Failure 404 {object} models.ErrorResponse "User not found"
// @Failure 500 {object} models.ErrorResponse "Internal server error"
// @Router /api/v1/users/{username} [get]
func GetUserDetailsHandler(c *gin.Context) {
	requestingUser := c.GetString("username")
	targetUser := c.Param("username")

	// Check authorization - must be superuser or requesting own account
	if !isSuperuser(requestingUser) && requestingUser != targetUser {
		log.Warnf("User '%s' attempted to access details for user '%s' without permission", requestingUser, targetUser)
		c.JSON(http.StatusForbidden, models.ErrorResponse{Error: "You don't have permission to access this user's details"})
		return
	}

	userDetails, err := auth.GetUserDetails(targetUser)
	if err != nil {
		// Check if user not found - using strings.Contains instead of exact match
		if strings.Contains(err.Error(), "unknown user "+targetUser) {
			log.Infof("User '%s' requested details for non-existent user '%s'", requestingUser, targetUser)
			c.JSON(http.StatusNotFound, models.ErrorResponse{Error: fmt.Sprintf("User '%s' not found", targetUser)})
			return
		}

		log.Errorf("Failed to get details for user '%s': %v", targetUser, err)
		c.JSON(http.StatusInternalServerError, models.ErrorResponse{Error: "Failed to retrieve user details: " + err.Error()})
		return
	}

	c.JSON(http.StatusOK, userDetails)
}

// @Summary Create user
// @Description Creates a new system user. Requires superuser privileges.
// @Tags Users
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param user body models.UserCreateRequest true "User creation details"
// @Success 201 {object} models.GenericSuccessResponse "User created successfully"
// @Failure 400 {object} models.ErrorResponse "Invalid request body"
// @Failure 401 {object} models.ErrorResponse "Unauthorized"
// @Failure 403 {object} models.ErrorResponse "Forbidden (User is not a superuser)"
// @Failure 409 {object} models.ErrorResponse "User already exists"
// @Failure 500 {object} models.ErrorResponse "Internal server error"
// @Router /api/v1/users [post]
func CreateUserHandler(c *gin.Context) {
	requestingUser := c.GetString("username")

	// Only superusers can create users
	if !requireSuperuser(c, requestingUser, "create a user") {
		return
	}

	var req models.UserCreateRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		log.Warnf("Invalid user creation request from user '%s': %v", requestingUser, err)
		c.JSON(http.StatusBadRequest, models.ErrorResponse{Error: "Invalid request body: " + err.Error()})
		return
	}

	// Create the user
	err := auth.CreateUser(req)
	if err != nil {
		// Check if the error indicates the user already exists
		if err.Error() == fmt.Sprintf("user '%s' already exists", req.Username) {
			log.Warnf("User '%s' attempted to create already existing user '%s'", requestingUser, req.Username)
			c.JSON(http.StatusConflict, models.ErrorResponse{Error: err.Error()})
			return
		}

		log.Errorf("Failed to create user '%s': %v", req.Username, err)
		c.JSON(http.StatusInternalServerError, models.ErrorResponse{Error: "Failed to create user: " + err.Error()})
		return
	}

	log.Infof("User '%s' successfully created new user '%s'", requestingUser, req.Username)
	c.JSON(http.StatusCreated, models.GenericSuccessResponse{Message: fmt.Sprintf("User '%s' created successfully", req.Username)})
}

// @Summary Update user
// @Description Updates an existing user. Regular users may update only their own profile fields. Group and superuser status changes require superuser privileges.
// @Tags Users
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param username path string true "Username to update" example="john"
// @Param user body models.UserUpdateRequest true "User details to update"
// @Success 200 {object} models.GenericSuccessResponse "User updated successfully"
// @Failure 400 {object} models.ErrorResponse "Invalid request body"
// @Failure 401 {object} models.ErrorResponse "Unauthorized"
// @Failure 403 {object} models.ErrorResponse "Forbidden (Not superuser or not the user's own account)"
// @Failure 404 {object} models.ErrorResponse "User not found"
// @Failure 500 {object} models.ErrorResponse "Internal server error"
// @Router /api/v1/users/{username} [put]
// internal/api/user_handlers.go - Fix for UpdateUserHandler
func UpdateUserHandler(c *gin.Context) {
	requestingUser := c.GetString("username")
	targetUser := c.Param("username")

	// Check authorization - must be superuser or requesting own account
	// Note: Regular users can't change superuser status even for themselves
	isRequestingSuperuser := isSuperuser(requestingUser)
	isSelfUpdate := requestingUser == targetUser

	if !isRequestingSuperuser && !isSelfUpdate {
		log.Warnf("User '%s' attempted to update user '%s' without permission", requestingUser, targetUser)
		c.JSON(http.StatusForbidden, models.ErrorResponse{Error: "You don't have permission to update this user"})
		return
	}

	var payload userUpdatePayload
	if err := c.ShouldBindJSON(&payload); err != nil {
		log.Warnf("Invalid user update request from user '%s': %v", requestingUser, err)
		c.JSON(http.StatusBadRequest, models.ErrorResponse{Error: "Invalid request body: " + err.Error()})
		return
	}

	req := payload.UserUpdateRequest

	// Only superusers may modify privilege-related fields.
	if !isRequestingSuperuser && (payload.hasGroups || payload.hasIsSuperuser) {
		log.Warnf("User '%s' attempted to modify privilege-related user fields", requestingUser)
		c.JSON(http.StatusForbidden, models.ErrorResponse{Error: "You don't have permission to modify user groups or superuser status"})
		return
	}

	// Update the user
	err := auth.UpdateUser(targetUser, req, isRequestingSuperuser)
	if err != nil {
		if errors.Is(err, auth.ErrPrivilegeUpdateForbidden) {
			log.Warnf("User '%s' attempted to update privilege-related fields for user '%s'", requestingUser, targetUser)
			c.JSON(http.StatusForbidden, models.ErrorResponse{Error: "You don't have permission to modify user groups or superuser status"})
			return
		}

		// Check if user not found - using strings.Contains instead of exact match
		if strings.Contains(err.Error(), "unknown user "+targetUser) {
			log.Infof("User '%s' attempted to update non-existent user '%s'", requestingUser, targetUser)
			c.JSON(http.StatusNotFound, models.ErrorResponse{Error: fmt.Sprintf("User '%s' not found", targetUser)})
			return
		}

		log.Errorf("Failed to update user '%s': %v", targetUser, err)
		c.JSON(http.StatusInternalServerError, models.ErrorResponse{Error: "Failed to update user: " + err.Error()})
		return
	}

	log.Infof("User '%s' successfully updated user '%s'", requestingUser, targetUser)
	c.JSON(http.StatusOK, models.GenericSuccessResponse{Message: fmt.Sprintf("User '%s' updated successfully", targetUser)})
}

// @Summary Delete user
// @Description Deletes a user from the system. Requires superuser privileges.
// @Tags Users
// @Security BearerAuth
// @Produce json
// @Param username path string true "Username to delete" example="john"
// @Success 200 {object} models.GenericSuccessResponse "User deleted successfully"
// @Failure 401 {object} models.ErrorResponse "Unauthorized"
// @Failure 403 {object} models.ErrorResponse "Forbidden (User is not a superuser)"
// @Failure 404 {object} models.ErrorResponse "User not found"
// @Failure 500 {object} models.ErrorResponse "Internal server error"
// @Router /api/v1/users/{username} [delete]
func DeleteUserHandler(c *gin.Context) {
	requestingUser := c.GetString("username")
	targetUser := c.Param("username")

	// Only superusers can delete users
	if !requireSuperuser(c, requestingUser, "delete user '"+targetUser+"'") {
		return
	}

	// Prevent deleting yourself
	if requestingUser == targetUser {
		log.Warnf("User '%s' attempted to delete their own account", requestingUser)
		c.JSON(http.StatusForbidden, models.ErrorResponse{Error: "You cannot delete your own account"})
		return
	}

	// Delete the user
	err := auth.DeleteUser(targetUser)
	if err != nil {
		// Check if user not found - using strings.Contains instead of exact match
		if strings.Contains(err.Error(), "unknown user "+targetUser) {
			log.Infof("User '%s' attempted to delete non-existent user '%s'", requestingUser, targetUser)
			c.JSON(http.StatusNotFound, models.ErrorResponse{Error: fmt.Sprintf("User '%s' not found", targetUser)})
			return
		}

		log.Errorf("Failed to delete user '%s': %v", targetUser, err)
		c.JSON(http.StatusInternalServerError, models.ErrorResponse{Error: "Failed to delete user: " + err.Error()})
		return
	}

	log.Infof("User '%s' successfully deleted user '%s'", requestingUser, targetUser)
	c.JSON(http.StatusOK, models.GenericSuccessResponse{Message: fmt.Sprintf("User '%s' deleted successfully", targetUser)})
}

// @Summary Change user password
// @Description Changes a user's password. Requires superuser privileges or the user's own account.
// @Tags Users
// @Security BearerAuth
// @Accept json
// @Produce json
// @Param username path string true "Username to change password for" example="john"
// @Param passwordChange body models.PasswordChangeRequest true "Password change details"
// @Success 200 {object} models.GenericSuccessResponse "Password changed successfully"
// @Failure 400 {object} models.ErrorResponse "Invalid request body"
// @Failure 401 {object} models.ErrorResponse "Unauthorized or incorrect current password"
// @Failure 403 {object} models.ErrorResponse "Forbidden (Not superuser or not the user's own account)"
// @Failure 404 {object} models.ErrorResponse "User not found"
// @Failure 500 {object} models.ErrorResponse "Internal server error"
// @Router /api/v1/users/{username}/password [put]
func ChangeUserPasswordHandler(c *gin.Context) {
	requestingUser := c.GetString("username")
	targetUser := c.Param("username")

	// Check authorization - must be superuser or requesting own account
	isRequestingSuperuser := isSuperuser(requestingUser)
	isSelfUpdate := requestingUser == targetUser

	if !isRequestingSuperuser && !isSelfUpdate {
		log.Warnf("User '%s' attempted to change password for user '%s' without permission", requestingUser, targetUser)
		c.JSON(http.StatusForbidden, models.ErrorResponse{Error: "You don't have permission to change this user's password"})
		return
	}

	var req models.PasswordChangeRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		log.Warnf("Invalid password change request from user '%s': %v", requestingUser, err)
		c.JSON(http.StatusBadRequest, models.ErrorResponse{Error: "Invalid request body: " + err.Error()})
		return
	}

	// Current password is required if not superuser
	if !isRequestingSuperuser && isSelfUpdate && req.CurrentPassword == "" {
		log.Warnf("User '%s' attempted to change their password without providing current password", requestingUser)
		c.JSON(http.StatusBadRequest, models.ErrorResponse{Error: "Current password is required"})
		return
	}

	// Change the password
	err := auth.ChangeUserPassword(targetUser, req.CurrentPassword, req.NewPassword, isRequestingSuperuser)
	if err != nil {
		// Handle specific error cases
		if err.Error() == fmt.Sprintf("user '%s' not found", targetUser) {
			log.Infof("User '%s' attempted to change password for non-existent user '%s'", requestingUser, targetUser)
			c.JSON(http.StatusNotFound, models.ErrorResponse{Error: err.Error()})
			return
		} else if err.Error() == "current password is incorrect" {
			log.Warnf("User '%s' provided incorrect current password while changing password", requestingUser)
			c.JSON(http.StatusUnauthorized, models.ErrorResponse{Error: err.Error()})
			return
		}

		log.Errorf("Failed to change password for user '%s': %v", targetUser, err)
		c.JSON(http.StatusInternalServerError, models.ErrorResponse{Error: "Failed to change password: " + err.Error()})
		return
	}

	log.Infof("User '%s' successfully changed password for user '%s'", requestingUser, targetUser)
	c.JSON(http.StatusOK, models.GenericSuccessResponse{Message: "Password changed successfully"})
}